+++ gromosXX/src/interaction/nonbonded/pairlist/standard_pairlist_algorithm.tcc
@@ -46,7 +46,7 @@ update(topology::Topology & topo,
        configuration::Configuration & conf,
        simulation::Simulation & sim, 
        Nonbonded_Set<t_interaction_spec, t_perturbation_spec> & nbs,
-       size_t begin, size_t end, size_t stride)
+       unsigned int begin, unsigned int end, unsigned int stride)
 {
   DEBUG(7, "standard pairlist update");
